+++ next.config.ts
@@ -1,8 +1,27 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
+  poweredByHeader: false,
   async headers() {
     return [
+      {
+        source: "/:path*",
+        headers: [
+          { key: "Referrer-Policy", value: "strict-origin-when-cross-origin" },
+          {
+            key: "Permissions-Policy",
+            value: "camera=(), geolocation=(), microphone=()",
+          },
+          { key: "X-Content-Type-Options", value: "nosniff" },
+          { key: "X-DNS-Prefetch-Control", value: "off" },
+          { key: "X-Download-Options", value: "noopen" },
+          { key: "X-Frame-Options", value: "SAMEORIGIN" },
+          {
+            key: "X-Permitted-Cross-Domain-Policies",
+            value: "none",
+          },
+        ],
+      },
       {
         source: "/charts/:path*",
         headers: [
@@ -23,6 +42,9 @@ const nextConfig: NextConfig = {
         pathname: "/images/**",
       },
     ],
+  },
+  turbopack: {
+    root: process.cwd(),
   },
 };